use matt's implementation and merge
