+++ src/renderer/components/defaults/default-in-list.js
@@ -3,21 +3,17 @@ import PropTypes from 'prop-types'
 
 import ContentTypes from '../../content-types'
 
-export default class DefaultInList extends React.PureComponent {
+export default class ListItem extends React.PureComponent {
   static propTypes = {
     url: PropTypes.string.isRequired,
-    hypermergeUrl: PropTypes.string.isRequired,
-    actions: PropTypes.arrayOf(PropTypes.string)
+    type: PropTypes.string.isRequired,
+    doc: PropTypes.shape({
+      title: PropTypes.string
+    })
   }
 
   static defaultProps = {
-    actions: []
-  }
-
-  state = {}
-
-  handleClick = (e) => {
-    window.location = this.props.url
+    doc: null
   }
 
   onDragStart = (e) => {
@@ -25,58 +21,37 @@ export default class DefaultInList extends React.PureComponent {
     e.dataTransfer.setDragImage(this.badgeRef, 0, 0)
   }
 
-  // This is the New Boilerplate
-  componentWillMount = () => this.refreshHandle(this.props.hypermergeUrl)
-  componentWillUnmount = () => this.handle.close()
-  componentDidUpdate = (prevProps, prevState, snapshot) => {
-    if (prevProps.hypermergeUrl !== this.props.hypermergeUrl) {
-      this.refreshHandle(this.props.hypermergeUrl)
-    }
+  render = () => {
+    const { doc, type } = this.props
+    const contentType = ContentTypes.lookup(type)
+
+    // TODO: this should be handled by the content system.
+    const icon = contentType ? contentType.icon : 'question'
+    const name = contentType ? contentType.name : 'Unknown'
+
+    // TODO: pick background color based on url
+    return (
+      <div style={css.listItem} draggable="true" onDragStart={this.onDragStart} className="DocLink" onClick={this.handleClick}>
+        <i ref={(ref) => { this.badgeRef = ref }} className={`Badge fa fa-${icon}`} />
+        <div className="DocLink__title">{(doc && doc.title) ? doc.title : name}</div>
+      </div>
+    )
   }
-
-  refreshHandle = (hypermergeUrl) => {
-    if (this.handle) {
-      this.handle.close()
-    }
-    this.handle = window.repo.watch(hypermergeUrl, (doc) => this.onChange(doc))
-  }
-
-  onChange = (doc) => {
-    this.setState({ ...doc })
-  }
-
-  renderActions = () => {
-    const actionOptions = []
-
-    if (this.props.actions.includes('archive')) {
-      actionOptions.push(<span key="archive" className="Type--secondary">⌘+⌫ Archive</span>)
-    }
-
-    if (this.props.actions.includes('view')) {
-      actionOptions.push(<span key="view" className="Type--secondary">⏎ View</span>)
-    }
-
-    if (actionOptions.length > 0) {
-      return <div className="Actions">{ actionOptions }</div>
-    }
-
-    return null
-  }
-
-  render = () => (
-    <div draggable="true" onDragStart={this.onDragStart} className="DocLink" onClick={this.handleClick}>
-      <i ref={(ref) => { this.badgeRef = ref }} className="Badge fa fa-files-o" style={{ background: this.state.backgroundColor }} />
-      <div className="DocLink__title">{ this.state.title }</div>
-      { this.renderActions() }
-    </div>
-  )
 }
 
+const css = {
+  listItem: {
+    padding: '5px',
+    border: '1px solid #eaeaea',
+    borderRadius: '4px'
+  }
+}
+
+
 ContentTypes.register({
-  component: DefaultInList,
-  type: 'default',
+  component: ListItem,
   context: 'list',
-  name: 'Document Link',
-  icon: 'sticky-note',
-  unlisted: true,
+  type: 'default',
+  name: 'Generic List Item',
+  icon: 'question'
 })
